restructure data import components
