+++ features/geojson/Sidebar/Data/index.tsx
@@ -1,10 +1,9 @@
 import dynamic from "next/dynamic";
 import { useStore } from "@geojson/store";
-import { ImportGeoJSON } from "./Import";
 import { Divider } from "@components/Divider";
+import { ImportModal } from "./ImportModal";
 
 const DataTable = dynamic(() => import("./DataTable"), { ssr: false })
-
 
 export function Data() {
   const geoJSONKey = useStore(state => state.geoJSONKey)
@@ -12,7 +11,7 @@ export function Data() {
 
   return (
     <>
-      <ImportGeoJSON />
+      <ImportModal />
       <Divider />
       {features.length !== 0 && <DataTable key={geoJSONKey} />}
     </>
